Fixing build

+++ core/services/ocr2/plugins/ccip/prices/exec_price_estimator_test.go
@@ -85,8 +85,8 @@ func TestExecPriceEstimator_GetGasPrice(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			sourceFeeEstimator := mocks.NewEvmFeeEstimator(t)
-			sourceFeeEstimator.On("GetFee", ctx, []byte(nil), uint32(0), assets.NewWei(tc.maxGasPrice)).Return(
-				tc.sourceFeeEstimatorRespFee, uint32(0), tc.sourceFeeEstimatorRespErr)
+			sourceFeeEstimator.On("GetFee", ctx, []byte(nil), uint64(0), assets.NewWei(tc.maxGasPrice)).Return(
+				tc.sourceFeeEstimatorRespFee, uint64(0), tc.sourceFeeEstimatorRespErr)
 
 			g := ExecGasPriceEstimator{
 				estimator:   sourceFeeEstimator,
